refactoring + change routing way + try fix netlify
